move all constants to separate file
